:truck: Rename package to bot
